Added explicit casts to pacify gcc.
git-svn-id: https://legesmotus.svn.sourceforge.net/svnroot/legesmotus/legesmotus/trunk@146 a5a5d5d7-22d1-44d1-b37a-67d3d99fcfd7

+++ common/Polygon.cpp
@@ -39,18 +39,18 @@ double	Polygon::intersects_circle(Point p, double radius) const {
 		int y1 = it->first.y;
 		int x2 = it->second.x;
 		int y2 = it->second.y;
-		double dtoline = fabs( (x2-x1)*(y1-p.y) - (x1-p.x)*(y2-y1) ) / sqrt( (x2-x1)*(x2-x1) + (y2-y1)*(y2-y1) );
-		double partofline = ((p.x - x1) * (x2 - x1) + (p.y - y1) * (y2 - y1)) / fabs(((x2-x1) * (x2-x1)) + ((y2-y1) * (y2-y1)));
+		double dtoline = fabs(double((x2-x1)*(y1-p.y) - (x1-p.x)*(y2-y1) ) / sqrt(double((x2-x1)*(x2-x1) + (y2-y1)*(y2-y1))));
+		double partofline = ((p.x - x1) * (x2 - x1) + (p.y - y1) * (y2 - y1)) / fabs(double(((x2-x1) * (x2-x1)) + ((y2-y1) * (y2-y1))));
 		if (dtoline < radius && partofline > 0 && partofline < 1) {
 			return dtoline;
 		}
 		
-		double dtocorner = sqrt((x1-p.x) * (x1-p.x) + (y1-p.y) * (y1-p.y));
+		double dtocorner = sqrt(double((x1-p.x) * (x1-p.x) + (y1-p.y) * (y1-p.y)));
 		if (dtocorner < radius) {
 			return dtocorner;
 		}
 		
-		dtocorner = sqrt((x2-p.x) * (x2-p.x) + (y2-p.y) * (y2-p.y));
+		dtocorner = sqrt(double((x2-p.x) * (x2-p.x) + (y2-p.y) * (y2-p.y)));
 		if (dtocorner < radius) {
 			return dtocorner;
 		}
